+++ gridstatus/tests/test_isos.py
@@ -204,7 +204,7 @@ def test_get_historical_load(iso):
                 "markets": [
                     Markets.REAL_TIME_HOURLY,
                     Markets.DAY_AHEAD_HOURLY,
-                    Markets.REAL_TIME_15_MIN,
+                    Markets.REAL_TIME_5_MIN,
                 ],
             },
         },
@@ -250,7 +250,7 @@ def test_get_historical_lmp(test):
                 "markets": [
                     Markets.REAL_TIME_HOURLY,
                     Markets.DAY_AHEAD_HOURLY,
-                    Markets.REAL_TIME_15_MIN,
+                    Markets.REAL_TIME_5_MIN,
                 ],
             },
         },
@@ -297,8 +297,8 @@ def test_get_latest_lmp(test):
         {
             CAISO(): {
                 "markets": [
+                    Markets.REAL_TIME_5_MIN,
                     Markets.REAL_TIME_HOURLY,
-                    Markets.REAL_TIME_15_MIN,
                     Markets.DAY_AHEAD_HOURLY,
                 ],
             },