+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## vNext 
 
 - Add CAISO LMP Heat Map Example Notebook
-- Add a guide on contributing to gridstatus
 - Add Settlement Point Prices for ERCOT
 - SPP: Add today/latest LMP for Real Time 5 minute and Day Ahead Hourly (DAM)
+- ERCOT load data is now returned with 5 minute frequency
+- Add a guide on contributing to gridstatus
 
 ## v0.16.0 - Dec 15, 2022